+++ Joker.Kafka/KSql/Query/Visitors/KSqlFunctionVisitor.cs
@@ -69,6 +69,7 @@ namespace Kafka.DotNet.ksqlDB.KSql.Query.Visitors
           case nameof(KSqlFunctionsExtensions.ArrayContains):
           case nameof(KSqlFunctionsExtensions.ArrayDistinct):
           case nameof(KSqlFunctionsExtensions.ArrayExcept):
+          case nameof(KSqlFunctionsExtensions.ArrayIntersect):
             Append($"{methodInfo.Name.ToKSqlFunctionName()}");
             PrintFunctionArguments(methodCallExpression.Arguments.Skip(1));
             break;