+++ Joker.Kafka/ChangeLog.md
@@ -1,8 +1,7 @@
-# v0.6.0-rc.1:
+# v0.6.0:
 - netstandard 2.0 (.Net Framework etc)
-- CASE - Select a condition from one or more expressions.
 - /Query endpoint (http 1.1)
-- Collection functions: None
+- CASE - Select a condition from one or more expressions.
 
 Added implementations:
 - QueryParameters, KSqlDbContextOptionsBuilder
@@ -114,6 +113,9 @@ Fixes:
 - KStreamSet, KQuerySet, KQueryStreamSet
 - QbservableProvider
 - ```KSqldbProvider<T>``` - ksqldb REST api provider for push queries (```KSqlDbQueryProvider<T>```, ```KSqlDbQueryStreamProvider<T>```)
+
+# v0.7.0 (WIP):
+- scalar collection functions: ArrayIntersect
 
 # TODO:
 - missing scalar functions https://docs.ksqldb.io/en/latest/developer-guide/ksqldb-reference/scalar-functions/#date-and-time